Problemas/Modulo #8/requirements.txt
Flask==2.0.1
redis==4.1.0
python-dotenv==0.19.1
celery==5.2.3
flask-mail==0.9.1